build views using runner callbacks
This reworks how views are rendered.  Instead of each view defining
a master template and a bunch of convoluted interactions with the
runner in order for things to work, this approach lets the runner
callback to the view at key points.  The view implements the
callbacks it cares about to render output.

available callbacks from the runner:
* `before_load`: called at the beginning, before the suite is loaded
* `after_load`: called after the suite is loaded, just before `on_start`, functionally equivalent to `on_start`
* `on_start`: called when a loaded test suite starts running
* `before_test`: called before a test starts running, the test is passed as an arg
* `on_result`: called when a running tests generates a result, the result is passed as an arg
* `after_test`:  called after a test finishes running, the test is passed as an arg
* `on_finish`:   called when the test suite is finished running

The callbacks are instance_eval'd in the view scope and have access
to any view methods.  Callbacks write output using the view's
`puts` and `print` methods (which work like Kernal#puts/print but
operate on the views `output_io`).