altered/added to the common helpers
Made the `result_details_for` now returns all results whether the
details for each should be shown or not.  It is now up to the view
using the helper to remove results whose details should not be
shown.

I also added a new helper `matched_result_details_for`.  This will
return result details much like `result_details_for` given a set
of tests.  However, it will only return tests that match some
criteria, either a file name or a context.  This is useful if
you want to group result output by either the file the test was
in or the context the test was in.