added a CLI (`assert`) for running tests with options
Rake is awful and I want to move away from supporting it.  This adds
a CLI to replace rake tasks for running tests.

The CLI globs any given paths with its test files and loads
the matching files.  This allows you to use tab-completion when
selecting which tests to run.  No test behavior has changed, just
the mechanism for loading multiple test files.

I added a deprecation warning to the rake task stuff.  They will be
removed in version 2.0.