updated Runner handling to configure runner instances
You now configure a Runner instance.  There are a couple reasons for
this.

One, the user may need to do custom Runner initialization and its
easier to let them do it and config Assert with an instance.  It's
never good to be creating instances of things on behalf of the users.

Two, this is more consistant with how you configure a View or a Suite -
you just pass an instance.