moved test file loading into the `Assert.init` method
Realized its probably better to let the `init` method control all
init behavior, including loading test files and firing view events.
This really isn't the business of the `AssertRunner` and before the
init behavior was divided.

This also updates the signature of the `init` method to take an
options hash.  This is for future behavior expansion.