handle and apply ENV settings
This updates the initialization to honor env var settings.  Right
now, runner_seed is the only setting driven by an env var.

I chose to remove the boolean options from the env settings b/c
of difficulty in parsing and the ambiguity in handling the values
parsed.  Ultimately it's easier to just use the flags anyway.  We
can always handle them at a later time.

Closes #98.
Closes #99.