don't hide backtraces on exceptions raised while loading the tests
Short-sightedly put in a feature where test loading exceptions had
their backtraces hidden except in debug mode.  This proved to be
very annoying as you always want to see the backtrace in this
scenario.  Plus, without showing the backtrace, I found myself not
recognizing the output as an error and it confused me greatly.

Closes #123.