have AssertRunner ensure test files are looked up from a list of paths
Before if no test_paths were provided from the CLI, the files were
looked up from a single `Assert.config.test_dir` path.  Magically,
the `inject` that worked on that single entry in MRI, but fails in
jruby and probably other ruby versions.

This just ensures that we are always injecting over a collection
of test paths when looking for test files to run.

Closes #125.