Inverting `show_output` option to be named `capture_output`
This is more intuitive and better matches the defaults.  The default
(capture_output = false) is still to show output as the tests are
run, but now the "easy" option `-o` does the opposite of the default
which is more appropriate.

Overall, makes this a more intuitive option while maintaining the
same default behavior.