add option for running only matching test files with changes
This is a new CLI option (`-c`) that tells assert to only load test
files in the given path(s) that have changes.  Use this to "scope
down" your test runs to just the tests you are actively working on.

This uses git (by default) to determine which files have changes.
Custom cmds can be configured in the user/local settings if git is
not available or a more specific set of cmds is needed.