show tests-per-sec and resuls-per-sec in views
This has the suite start calculating the rate of tests and results
so that the view can have access to the it.  It also adds it to the
summary of the default view.

This is just so the user has extra stats to keep an eye on their
test suite run time.  It provides a more granular measurement on
the changes in suite run time.  Plus it is fun.