only display ANSI styled output if the view IO is a tty
This updates the view to know whether it is a tty or not and only
display ansi styled output if it is a tty - regardless of any other
config settings.  This is to be a good citizen and not output ANSI
style coded output to IOs that don't know what to do with it.