cleanups/standardizations to assertion fail messages
This standardizes all assertion fail messages to use the format:
`"Expected {act-value} to {not }be {something|exp-value}".

Before, the messages had difference formats depending on various
things like the nature of the assertion, when the assertion was
added, and whether it was a "not" assertion.  This cleans things
up and makes them consistent so there is no guessing how to write
the fail messages when implementing a new assertion.

This also includes some cleanups to the assertion tests.