capture individual test run time and result rate
Much like the suite, the captures run time and result rate for
individual tests.  More stats the better, right?  This will be used
in an upcoming feature to display test details and profile info in
the default view.

I choose this benchmark implementation b/c it had the least amount
of performance impact.