add option to show test profile info; add to the default view
This adds a new cli option, `-e`, to switch whether to show test
profile info.  This also adds some common view methods for querying
and displaying profile info.  The tests are outputted from fastest
run time to slowest so the slowest tests are the most obvious.

Finally, this adds displaying this profile info is that option is on
to the default view.  In the default view, "profile info" means
the test run time, test result count, test result rate, plus the test
class/name for each test.