Add readable inspect to `Stub`
This adds a readable inspect to `Stub` to help with cleaner
output. Since asserts stubbing now writes instance variables, the
default inspect for objects will show the stubs that have been
written to an object. This would inspect the instances of `Stub`
on the object which would show all of its instance variables. The
output of this was noisy and difficult to read. This cleans up
the output and makes it simpler to read.