verbose output mode
This adds a verbose runtime output mode flag to the CLI and config.
This also adds a method for querying this flag to the common view
helpers.  Finally, this updates the default view to honor the
flag in its output.

This goal here is to be able to give more details about what is
running at runtime.