track the file and line number for each test
Using the context info's called from attribute, we can pull out
and track which file and line number a test is defined on.  This
will allow running tests by file/line-number and is prep for that
feature.

Note: this also updates the methods macro to properly pass called from
context to its test definition.  The single methods test will be
seen a defined on the line of the first methods macro in the test
file.  I noticed this was incorrect while debugging for the file
line number change.

Note also: this updates assert's own factory to extend assert factory.
This opens up all the factory methods for use in assert's own test
suite.  Note really sure why this wasn't done already - it was
probably just missed.  I switched to using factory methods to build
the context info called from factory.