dump test results on interrupt
This updates the runner to catch `Interrupt` exceptions and to fire
a new view callback `on_interrupt` before re-raising the interrupt.
This allows view handlers to handle the interrupt scenario.

This also updates the default view to dump its test results on
interrupt just like it would on finish.  This allows you to
interrupt a long test suite run that is producing a bunch of errors
and see what those errors are before having to wait for the test
suite run to finish.