tweak runner interrupt handling to not include view start/finish events
This is to make the interrupt handling a bit more focused.  There
is nothing to interrupt if the suite hasn't been setup yet.  Also
there is nothing to interrupt if the suite has already been torn
down.

Practically, in the default view, this will prevent the case where
you interrupt while the on finish event is dumping results, only
to then dump those results again.

/cc @jcredding