`FileLine` helper class for file/line-number handling
This adds a helper class for dealing with test file/line-numbers.
It includes the basic readers but also defines how these are
represented in strings and also how to parse from a string.  It
also encapsulates comparison logic.

In addition, this defines an `AnyFileLine` utility class.  This
class is special in that it will be equal to any file line it is
compared with.

`Test` has been backported to track its file line.  This will be
used in the coming run-by-line-number feature.  File lines will
be taken as args from the cli and only run tests with matching file
lines will be run.  This will also be used on the coming verbose
output feature as we will display the test's file line in the
output.