`-l` option for listing out test files that would be loaded and run
Use this option to list out the test files the *would be* loaded
and run to stdout.  There are two goals here: 1) use this to help
debug and 2) use in a script to run each test file individually.

Note: this reorganizes the CLI options to list in a more relevent
order.  This also reorganizes the config option settings to group
them together and order them in the same order as the CLI defines
them.  Just cleanups and readability improvements here.