move common view helpers into top level file; test them
This cleans up the common view helpers to organize them more like
the new config helpers.  This also adds a proper set of tests for
them instead of just relying on the test views not breaking to test
them.

This is just a cleanup to make things more robust - there are no
behavior changes.

This did move the helper that gets the occurring result types to
the more general config helpers as it relies on the `count` method
provided by them and is less about presentation and more about just
general state.

This also moves the helper that would get the ordered tests for
profile output to the suite and renames it more by what it does:
sort the ordered tests by run time.  This didn't feel like a view
helper as much as a set of data that should be provided by the
suite.  The weird helper name is the "code smell" that initiated
this move.

Finally, I had to update the result factories to do the view helper
tests.  I provided the additional missing ones and simplified the
api on the existing skip result factory.