have runner describe how tests are going to be run
Before, the runner just fired events on the view and the view
could determine what to say on start.  However, the default view
would make assumptions about the runner and say that we are running
tests in random order when in fact this may not be true.

This moves the logic of describing the run into the runner as it
isn't a view's concern.  This is prep for intending the runner to
be extended - different runners may want to both use the default view
but may run their tests in different manners.

Note: the view continues to have no control on the order that output is
displayed - the runner controls when events are fired which controls
when the different output pieces are displayed.