remove the `ansi` gem as a dependency
The goal here is to make assert have no dependencies (your test
framework shouldn't add a bunch of dependencies to your app).  In
addition assert only needs a fraction of the abilities ansi provided.

This reworks the ansi style code handling to not require the ansi
gem and to only implement the specific api needed by assert.