simplify, standardize view constructor
This removes the ability to dynamically specify view configs/suites.
This ability isn't really serving any purpose and is an artifact
of legacy implementations.  This is prep for Parassert - making
the core objects more intentionally extendable.

I also chose to change it so that you pass the config as the first
arg to the view constructor.  This makes is consistent with the
suite and runner objects.  I'm trying to centralize all handling
around the config object and locking things down so that the config
truly drives all settings.  #221 and #222 started this effort by
centralizing the helpers around the config object.