simplify runner run method; remove suite/view injection
This is similar to 226 and is prep for doing Parassert.  The goal
here is to make the runner intentionally extendable yet as simple
to work with as possible.  This also continues the efforts of 221
and 222 to make the config the common central point for all settings.

The `run` methods dependency injection is an artifact of previous
implementations and currently serves no purpose.  It only adds
complexity.