have results capture their data in a specific "data" objects
This is similar to the recent update to Tests where they caputure
their raw data in data objects.  The goal here is to allow more
flexible workflows with result data - serializing just the data
bits and deserializing them for example.  Parassert will use the
to store result data in a db and rebuild from the db when outputting
results.

Note: this also includes some updates to the test system tests.  They
were relying on monkeying with test/result internals to test their
logic.  This is bad form and brittle and was shown to be when the
tests broke on this change.  I've updated the tests to use non-internal
logic and more safely implement their tests.