remove `fire` method on the base view
Remove the ceremony of going through a custom dispatch method.  Just
call the methods directly.  There was nothing wrong with the previous
implementation, I'm just trying to simplify things.  This also updates
the callbacks comment to make it more accurate.

This is prep for adding a similar callback setup for the suite and
the runner.