rework suite implementation/handling; add DefaultSuite
This reworks the suite object to make it more flexible to extend.
This follows the pattern of the view object where the base suite
provides the core logic/behavior and the default suite provides
the default assert behavior.  The goal here is to enable extending
the suite for use in AssertSqlite (and Parassert).

The suite is responsible for accumulating tests for the runner to
run and for providing the run data and test result data.  The default
suite stores run data and results in memory.

Note: this begins the work of formalizing the test/result data api.
In a coming effort, I'll rework the default view to fully use this
api and implement the full api in the default suite.

There are a few other smaller changes here:

* I broke out the context info class into its own top-level object.
  This is purely an organizational change as it has nothing to do
  with the suite object.
* I simplified how the setups/teardowns are added and accessed.  This
  removes the unecessary complexity of the previous implementation.
* I added a callback setup similar to that of the view.  The goal
  here is to allow custom suites to inject special behavior during
  test runs (much like the views can).  This is just to make suites
  as extendable as possible and provide the same flexibility as the
  views.