update view to no longer use a base class
This is a purely organizational change.  The goal here is to make
the view implementation more closely match the new suite implementation
and to remove some legacy ceremony/complexity with view handling.

The diff is rugged b/c there was a tab indention change - there are
no behavior changes here.