update view to get all test/result data from the suite
Before the view gathered its data partially from the suite and
partially processed it on its own.  This would break if the suite
implements custom ways of storing and making data available.

This reworks the views and view helpers to get all data from the
suite and updates the default suite to implement all suite behavior
by storing data in memory and processing it in memory.

This is prep for assert-sqlite where its suite will not store data
in memory but will instead store it in the db and query the db for
data processing.

There are a few other tweaks I've added here:

* added `output` to the result api and handling - it is needed by
  the views and should have been added originally
* switched from array `count` to array `size` - supposedly it is
  faster: https://github.com/JuanitoFatas/fast-ruby#arraylength-vs-arraysize-vs-arraycount-code
* removed a bunch of view helpers that are seemingly not used.  They
  may have been for kellyredding/assert-view-leftright and
  kellyredding/assert-view-testunit - both of which are deprecated.
* shortened the "reverse_*" api methods for less typing