hotfix: update on finish handling to allow suite/runner cleanup
We need to run the view on finish first b/c it most likely needs
suite data.  We should run the runner's callback last to be in true
reverse order of the on start logic.  This is all done in a tap
so that we can still return the fail/error result counts after the
suite on finish callback has run.

This will especially help assert-sqlite, who needs to drop the
db in its on finish callback.

Not: I also switched to the more standard `:fail` and `:error`
result type args for the counting logic.  Just trying to be more
standard - no technical reason.

/cc @jcredding