backfill Array#sample so tests pass in ruby 1.8.7
The test suite was updated to use `Array#sample` instead of
`Array#choice` in 239 so that the test suite would pass in modern
ruby versions.  However, I would prefer that the test suite continues
to pass in ruby 1.8.7 as well.

This backfills the `sample` method so that tests will pass in 1.8.7.