accumulate: prep notes for accumulating test results
I made a bunch of notes of things to change when switching to
accumulating test results.  Currently we store every test object
and every result generated by every test.  This is inefficient
for calculating summary info at the end of the test run and is
unnecessary as we don't need to store all of this data.  Plus
the overhead on large test suites is significant.

In this overall effort, I plan on switching to not storing every
test.  Instead, tests will be shifted off as they are run and
only result data needed for presentation will be accumulated.
This should save memory not storing as many objects and should
make rendering summary presentation data more efficient.